Move array shape routines to their own utils module.
