+++ astropy/_erfa/core.py.templ
@@ -35,7 +35,7 @@ import numpy
 # we import these exceptions from astropy locations instead of defining them
 # in this file because otherwise there are circular dependencies
 from astropy.utils.exceptions import ErfaError, ErfaWarning
-from astropy.utils.misc import check_broadcast
+from astropy.utils import check_broadcast
 
 from . import ufunc
